Add WHIRLPOOL implementation
Implement the WHIRLPOOL hashing function (version 3.0), tested against
the ISO test vectors (provided as part of ISO/IEC 10118-3:2004 and
included with the WHIRLPOOL reference implementation).

Hashing an arbitrary amount of bits is not yet supported.

This implementation slightly diverges from WHIRLPOOL by limiting the
maximum size of the input to `2^64 - 1` bits (~2.1 million terabytes).